Adaptação do AddReceitas para o filtro da geladeira

+++ Tempera/SalvarGeladeira.php
@@ -1,6 +1,11 @@
 <?php 
- session_start();
-
+session_start();
+if(!isset($_SESSION['id_usuario']))
+{
+    header("location: Index.php");
+    exit;
+    
+};
 
  $link = mysqli_connect("clovis-cartola.czcbeh0esbig.us-east-1.rds.amazonaws.com", "tempera", "Tempera_123", "tempera");
 
@@ -12,7 +17,7 @@ if (!$link) {
     echo "Debugging error: " . mysqli_connect_error() . PHP_EOL;
     exit;
 }else {
-        print_r($_POST);
+    if($_GET['Funcao'] == 'Salvar'){
         $id = $_SESSION['id_usuario'];
         $query = "delete from tb_geladeira where id_usuario ={$id}";
         mysqli_query($link,$query);
@@ -21,7 +26,16 @@ if (!$link) {
             $addIngre = "insert into tb_geladeira (id_usuario, st_nomeIngrediente) values ('{$id}','{$ingre}')";
             mysqli_query($link,$addIngre);           
         };
-       header('location:Geladeira.php');
+    }else if($_GET['Funcao'] == 'Filtrar'){
+        $url = 'location:Home.php?';
+        foreach($_POST as $Ingre){
+            $url = $url.$Ingre."=".$Ingre."&";
+        };
+        header($url);            
+    }else{
+        header('location:Geladeira.php');
+    }
+   //header('location:Geladeira.php');
 }
 
 ?>
